pyproject.toml
[tool.poetry]
name = "dj"
version = "0.1.0"
description = ""
authors = ["Adam Hill <adamghill@yahoo.com>"]
include = ["dj-config.json"]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
"delegator.py" = "^0.1.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
dj = 'dj:main.run'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
